refactoring.  removing height prop
